Authentication and Authorization

+++ controllers/api/v1/posts_api.js
@@ -23,14 +23,19 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
-    // if (post.user == req.user.id) {
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-      message: 'Post and associated comment delete succesfully!',
-    });
+      return res.json(200, {
+        message: 'Post and associated comment delete succesfully!',
+      });
+    } else {
+      return res.json(401, {
+        message: 'You cannot delete this post!',
+      });
+    }
   } catch (err) {
     return res.json(500, {
       message: 'Internal Server Error',
